requirements.txt
aiohttp==3.7.3
python-decouple==3.4
2captcha-python==1.0.3
Scrapy==2.4.1
pydantic==1.7.3
pytz==2021.1